requirements.txt
python-nmap
pyfiglet
wfuzz

